test(workflow): move chat-wiring tests into validation_taskmode_test.go
Fold the chat-mode wiring tests into the existing mode-validation test
file instead of a separate file; reuses the sibling's node-builder idiom.
